requirements.txt
python_dateutil >= 2.5.3
setuptools >= 21.0.0
urllib3 >= 1.25.3, < 2.1.0
pydantic >= 2
typing-extensions >= 4.7.1
PyNaCl==1.5.0
fastapi==0.103.2
uvicorn==0.22.0
ecdsa==0.19.0
